Update index.html
In this project, a website for “BLACK MEAT” food ordering was created. Various food categories were added, each dish includes an image, description, price, and an order button. The homepage features the restaurant’s introduction and services. A filtering function by menu categories was implemented, and a navigation menu was set up.